refactor(stores): decompose useUnitStore into composed sub-stores
- useUnitStore.ts: 1,280 → 18 lines (re-export shim for backwards compatibility)
- Created src/stores/unit/ with 5 files:
  - useUnitStore.ts (140L) - composition shell combining slices
  - useUnitArmorStore.ts (155L) - armor allocation actions
  - useUnitEquipmentStore.ts (261L) - equipment add/remove/mount
  - useUnitStructureStore.ts (589L) - engine/gyro/structure with cascade
  - useUnitTechBaseStore.ts (274L) - tech base mode switching
- Equipment cascade logic centralized in structure slice
- Persistence key megamek-unit-${id} preserved
- All 73 characterization tests pass
- All 957 store tests pass
- UnitStoreProvider contract preserved (no consumer changes needed)
